% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{aeso.data}
\alias{aeso.data}
\alias{SJSyndrome.data}
\alias{MySeqList}
\alias{dico}
\alias{LongSeq}
\alias{MidSeq}
\alias{ShortSeq}
\title{Deprecated}
\format{
Refers to the "See also" links below to obtain a description of each dataset.
}
\usage{
data(Aeso) ; data(SJSyndrome) ;
       data(SeqListSCOPe) ; data(HydroScore) ;
       data(Seq1093) ; data(Seq219) ; data(Seq31)
}
\description{
Some datasets are deprecated and will be removed in next version of the package.
Please use instead:
\itemize{
  \item data(aeso.data) -> data(Aeso)
  \item data(SJSyndrome.data) -> data(SJSyndrome)
  \item data(MySeqList) -> data(SeqListSCOPe)
  \item data(dico) -> data(HydroScore)
  \item data(LongSeq) -> data(Seq1093)
  \item data(MidSeq) -> data(Seq219)
  \item data(ShortSeq) -> data(Seq31)
}
}
\seealso{
\code{\link{Aeso}} \code{\link{SJSyndrome}} \code{\link{SeqListSCOPe}} 
         \code{\link{HydroScore}} \code{\link{Seq1093}} \code{\link{Seq219}} \code{\link{Seq31}}
}
\keyword{internal}
