% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rMethods.R
\name{CharSequence2ScoreSequence}
\alias{CharSequence2ScoreSequence}
\title{Convert a character sequence into a score sequence}
\usage{
CharSequence2ScoreSequence(sequence, dictionary)
}
\arguments{
\item{sequence}{a character sequence}

\item{dictionary}{a dictionary}
}
\value{
a vector of a score sequence
}
\description{
Convert a character sequence into a score sequence. See CharSequences2ScoreSequences() fonction for several sequences
}
\examples{
data(ShortSeq)
ShortSeq
data(dico)
CharSequence2ScoreSequence(ShortSeq,dico)
}
