% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rMethods.R
\name{CharSequences2ScoreSequences}
\alias{CharSequences2ScoreSequences}
\title{Convert several character sequences into score sequences}
\usage{
CharSequences2ScoreSequences(sequences, dictionary)
}
\arguments{
\item{sequences}{a list of character sequences}

\item{dictionary}{a dictionary}
}
\value{
a list of score sequences
}
\description{
Convert several character sequence into score sequences. For only one sequence see CharSequence2ScoreSequence() function.
}
\examples{
data(ShortSeq)
ShortSeq
data(MidSeq)
MidSeq
data(dico)
MySequences=list("A1"=ShortSeq,"A2"=MidSeq)
CharSequences2ScoreSequences(MySequences,dico)
}
