% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dico}
\alias{dico}
\title{Deprecated. Use 'HydroScore' instead.}
\format{
Deprecated. Use 'HydroScore' instead.
}
\usage{
dico
}
\description{
Deprecated. Use 'HydroScore' instead.
}
\keyword{datasets}
