% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rMethods.R
\name{loadMatrixFromFile}
\alias{loadMatrixFromFile}
\title{Loads matrix from csv-File}
\usage{
loadMatrixFromFile(filepath)
}
\arguments{
\item{filepath}{optional: Location of file on disk. If not provided, a file picker dialog will be opened.}
}
\value{
A Matrix Object
}
\description{
Reads a csv file without header and returns the matrix. For file formats please see section "File Formats" in vignette.
}
