\name{localScore-package}
\alias{localScore-package}
\alias{localScore}
\docType{package}
\title{
  Package for sequence analysis by local score
}
\description{
  Provides functionalities for: 
  \itemize{
      \item calculating the local score 
      \item calculating statistical relevance (p-value) to find a local Score in a sequence of given distribution. 
  }
}
\details{
  Please refer to the vignette of this package or the manual for details on how to use this package.
}
\author{
Sebastian Simon, Sabine Mercier, David Robelin, Anne-Benedicte Urbano

Maintainer: David Robelin david.robelin@inra.fr
}
\references{
  p-value:
    \itemize{
      %\item Exact Distribution of the Local Score for Markovian Sequences, Claudie Hassenforder and Sabine Mercier, 2006 
      \item An Improved Approximation For Assessing The Statistical Significance of molecular Sequence Features, Mercier and al 2003  
      \item Exact distribution for the local score of one i.i.d. random sequence, Sabine Mercier and JJ Daudin, 2001
      \item Limit Distributions of Maximal Segmental Score among Markov-Dependent Partial Sums, Karlin and Dembo 1992
      \item Methods for assessing the statistical significance of molecular sequence features by using general scoring schemes, Karlin and al 1990
  }
  local Score:
  \itemize{
    \item Detection de courts segments inverses dans les genomes: methodes et applications, David Robelin 2005
    \item A Linear Time Algorithm for Finding All Maximal Scoring Subsequences, Constable and Bates 1985
  }
}
