% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excursion.R
\name{MarkovParameters}
\alias{MarkovParameters}
\title{Functions parameters global description}
\arguments{
\item{theta}{alphabet used (vector of character)}

\item{lambda}{transition probability matrix of size, theta x theta}

\item{score_function}{vector containing the scores of each letters of the alphabet (must be in the same order as theta)}

\item{a}{score strictly positive}

\item{prob0}{Distribution of the first element of the Markov chain. Default to stationary distribution of lambda}
}
\description{
Functions parameters global description
}
