% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MySeqList}
\alias{MySeqList}
\title{Deprecated. Use 'SeqListSCOPe' instead.}
\format{
Deprecated. Use 'SeqListSCOPe' instead.
}
\usage{
MySeqList
}
\description{
Deprecated. Use 'SeqListSCOPe' instead.
}
\keyword{datasets}
