% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rMethods.R
\name{scoreDictionnary2probabilityVector}
\alias{scoreDictionnary2probabilityVector}
\title{Check for missing scores values in the score distribution}
\usage{
scoreDictionnary2probabilityVector(list, score_extremes)
}
\arguments{
\item{list}{vector of scores}

\item{score_extremes}{vector of probability}
}
\value{
vector containing all score values between extremes and the probability equal to 0 for missing score
}
\description{
Get extremes scores, then create a complete list and set a probability equal to zeros for not present scores
}
\examples{
Mylist=list("x1"=c(-2,0.1),"x2"=c(0,0.7),"x3"=c(1,0.2))
scoreDictionnary2probabilityVector(list=Mylist,score_extremes=c(-2,1))
}
