% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Seq1093}
\alias{Seq1093}
\title{Long protein sequence}
\format{
A character string with 1093 characters corresponding to Q60519.fasta in UniProt Data base.
}
\source{
\url{https://www.uniprot.org/}
}
\usage{
data(Seq1093)
}
\description{
A long protein sequence.
}
\examples{
data(Seq1093)
Seq1093
nchar(Seq1093)
data(HydroScore)
seqScore <- CharSequence2ScoreSequence(Seq1093,HydroScore)
seqScore[1:50]
localScoreC(seqScore)$localScore
LS <- localScoreC(seqScore)$localScore[1]
prob1 <- scoreSequences2probabilityVector(list(seqScore))
daudin(local_score = LS, sequence_length = nchar(Seq1093),
       score_probabilities = prob1,
       sequence_min = min(seqScore),
       sequence_max = max(seqScore))
karlin(local_score = LS, sequence_length = nchar(Seq1093),
       score_probabilities = prob1,
       sequence_min = min(seqScore),
       sequence_max = max(seqScore))
}
\keyword{datasets}
