% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SJSyndrome}
\alias{SJSyndrome}
\title{Stevens-Johnson syndrome data}
\format{
A data.frame of 824 lines, each describing a syndrome appearance described by 15 covariates:
Case ID,
Initial FDA Received Date,
days since last fda,
Event Date,
Latest FDA Received Date,
Suspect Product Names,
Suspect Product Active Ingredients,
Reason for Use,
Reactions,
Serious,
Outcomes,
Sex,
Patient Age,
Sender,
Concomitant Product Names
The third column correspond to the number of days between two adverse events.
}
\source{
FDA open data.
}
\usage{
data(SJSyndrome.data)
}
\description{
The Stevens-Johnson syndrome is an acute and serious dermatological disease due to a drug allergy. 
The syndrome appearance is life-threatening emergency. They are very rare, around 2 cases per million people per year.
}
\examples{
data(SJSyndrome)
summary(SJSyndrome)

}
\keyword{datasets}
