% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{stationary_distribution}
\alias{stationary_distribution}
\title{Stationary distribution [Markov chains]}
\usage{
stationary_distribution(m)
}
\arguments{
\item{m}{Transition Matrix [matrix object]}
}
\value{
A vector with the probabilities
}
\description{
Calculates stationary distribution of markov transitition matrix by use of eigenvectors of length 1
}
\examples{
B = t(matrix (c(0.2, 0.8, 0.4, 0.6), nrow = 2))
stationary_distribution(B)
}
