% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{aeso.data}
\alias{aeso.data}
\title{Deprecated. Use 'Aeso' instead.}
\format{
Deprecated. Use 'Aeso' instead.
}
\usage{
aeso.data
}
\description{
Deprecated. Use 'Aeso' instead.
}
\keyword{datasets}
