% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rMethods.R
\name{CharSequence2ScoreSequence}
\alias{CharSequence2ScoreSequence}
\title{Convert a character sequence into a score sequence}
\usage{
CharSequence2ScoreSequence(charseq, dictionary)
}
\arguments{
\item{charseq}{a character sequence, given as a string}

\item{dictionary}{a data.frame with rownames containing letters, first column containing associated scores, optional second column containing associated probabilities}
}
\value{
a vector of a score sequence
}
\description{
Convert a character sequence into a score sequence. See CharSequences2ScoreSequences() function for several sequences
}
\examples{
data(Seq31)
Seq31
data(HydroScore)
CharSequence2ScoreSequence(Seq31,HydroScore)
}
\seealso{
\code{\link{CharSequences2ScoreSequences}}
}
