% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rMethods.R
\name{CharSequences2ScoreSequences}
\alias{CharSequences2ScoreSequences}
\title{Convert several character sequences into score sequences}
\usage{
CharSequences2ScoreSequences(sequences, dictionary)
}
\arguments{
\item{sequences}{a list of character sequences given as string}

\item{dictionary}{a data.frame with rownames containing letters, first column containing associated scores, optional second column containing associated probabilities}
}
\value{
a list of score sequences
}
\description{
Convert several character sequence into score sequences. For only one sequence see CharSequence2ScoreSequence() function.
}
\examples{
data(Seq31)
Seq31
data(Seq219)
Seq219
data(HydroScore)
MySequences=list("A1"=Seq31,"A2"=Seq219)
CharSequences2ScoreSequences(MySequences,HydroScore)
}
\seealso{
\code{\link{CharSequence2ScoreSequence}}
}
