% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rMethods.R
\name{loadScoreFromFile}
\alias{loadScoreFromFile}
\title{Load score from file}
\usage{
loadScoreFromFile(filepath, ...)
}
\arguments{
\item{filepath}{optional: location of file on disk. If not provided, a file picker dialog will be opened.}

\item{...}{optional: use arguments from read.csv}
}
\value{
A List Object - Names correspond to the first column, usually Letters. Associated numerical scores are in the second column. If probabilities are provided, 
they will be loaded too and presumed to be in the third column
}
\description{
Reads a csv file with 2-3 columns and returns it as a list object of vectors, with names corresponding to the first column of the file. For details view section "File Formats" in vignette.
}
