% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{LongSeq}
\alias{LongSeq}
\title{Deprecated. Use 'Seq1093' instead.}
\format{
Deprecated. Use 'Seq1093' instead.
}
\usage{
LongSeq
}
\description{
Deprecated. Use 'Seq1093' instead.
}
\keyword{datasets}
