% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MidSeq}
\alias{MidSeq}
\title{Deprecated. Use 'Seq219' instead.}
\format{
Deprecated. Use 'Seq219' instead.
}
\usage{
MidSeq
}
\description{
Deprecated. Use 'Seq219' instead.
}
\keyword{datasets}
