% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ShortSeq}
\alias{ShortSeq}
\title{Deprecated. Use 'Seq31' instead.}
\format{
Deprecated. Use 'Seq31' instead.
}
\usage{
ShortSeq
}
\description{
Deprecated. Use 'Seq31' instead.
}
\keyword{datasets}
